backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prism-backend"
version = "0.1.0"
description = "PRISM FastAPI backend for CIP/CASVE career decision research"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy[asyncio]>=2.0.35",
  "asyncpg>=0.30.0",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.16.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "python-dotenv>=1.0.1",
  "pgvector>=0.3.6",
  "openai>=1.99.0",
  "tavily-python>=0.7.0",
  "httpx>=0.28.0",
  "orjson>=3.10.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.0",
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
markers = [
  "live: end-to-end tests against real OpenAI/Tavily services",
]

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "I", "UP", "B"]
